Bring over erl_parse from erl2 as erlfmt_parse
Summary:
This simply copies over the erl_parse module from erl2 as erlfmt_parse.
Further diffs will modify the parser to support compilation of macros without expansion and remove what is not needed for the formatter to reduce the scope area.
Having a separate parser module gives us more leeway in changes we can make and is less critical for issues or possible bugs introduced.