Change record AST in erlfmt
Summary:
This diff changes the representation of records in the `erlfmt_parse` parser, in particular the
record name is represented differently.
The new representation is the same in all the places: types, expressions and patterns. Previously,
the representation was different in types and in code.
The new representation for the name is always `{atom, _, Name}` instead of a raw atom `Name`.

By consequence this drops the module record syntax from the erlfmt parser. When needed,
it should be possible to add it failrly easily.

This representation has two main advantages:
  * it unifies the representation between types and code
  * it preserves maximum amount of annotations that will be important for the formatter.

To support this change, both the parser and the associated typespecs are changed.
Ann additional documentation document listing all the differences between the erlfmt parser and
the erl1 parser is created.

Test Plan: `rebar3 ct erlfmt`

Reviewers: alavrik, #erl2

Reviewed By: alavrik

Subscribers: ilyaklyuchnikov, alavrik, viz

Differential Revision: https://phabricator.intern.facebook.com/D17855315

Tasks: T54121019

Tags: erl2

Signature: 17855315:1570799687:44d9b869afc07701ed2f6a328e4e0ec0bcc6f014