Missing tests and docs for erlfmt
Summary:
Adds tests for the record AST changes introduced in D17855315,
and documents and tests the attribute AST changes introduced in D17792024.

The reason for the original change of attribute AST was to preserve metadata
in the values of attributes. The original parser would completely drop any annotations
and only retain concrete values in the AST