Make function AST more flexible
Summary:
This prepares the `function` and `fun` AST nodes to accomodate changes required
for parsing macros. In the following diffs, various elemenents that were changed
from raw atoms to full atom nodes will be allowed to also hold other kind of nodes.

Additionally, we might not always have a consistent name and arity for a function,
for example where some of the clauses are defined using a macro. Dropping the fields
is the only reasonable option.