Introduce macro_call AST node
Summary:
This introduces a new `macro_call` AST used for representing unexpanded macro
calls in expressions.

The `macro_call_none` rule is separated as it will be useful in more places, where
full macro calls would be problematic.

A parsing shift/reduce conflict is introduced because of the optional parenthesis
for macro calls. It is resolved correctly by the default yecc rules.

> Parse action conflict scanning symbol '(' in state 240:
> Reduce to macro_call_none from '?' atom_or_var (rule 241 at line 507)
> vs.
> shift to state 243, adding right sisters to atom_or_var.
> Conflict resolved in favor of shift.