Make remote a standalone expression
Summary:
This changes the `remote` node to also be a standalone expression.
This is done primarily to support macros as in the added test.
Additionally, this is returning to how it was handled in the original erl_parse parser.

Test Plan: `rebar3 ct erlfmt`

Reviewers: ilyaklyuchnikov, #erl2

Reviewed By: ilyaklyuchnikov

Differential Revision: https://phabricator.intern.facebook.com/D17929799

Tasks: T54121019

Tags: erl2

Signature: 17929799:1571246205:e24e6958f649f45c100757693aa155524f83ccad