Remove attribute value conversion
Summary:
This removes the special-cased conversion that was done on atrribute
values. In particular the parser would convert the `fun/arity` syntax in
attributes from applications of the `/` operator into 2-tuples. This
dropped a lot of metadata. If we'll need to print those differently
than regular applications of the `/` operator, it will be better to
special-case those in the printer rather than the parser.

This also allows for other nodes than just `atom` to be used as
function name in `export` and `import` attributes - `macro_call` nodes
in particular.

Test Plan: `rebar3 ct erlfmt`

Reviewers: ilyaklyuchnikov, #erl2

Reviewed By: ilyaklyuchnikov

Differential Revision: https://phabricator.intern.facebook.com/D17931925

Tasks: T54121019

Tags: erl2

Signature: 17931925:1571301642:07eff62b3283863b4b3c41f270603b2f0ec7987a