Introduce lines API to erlfmt_algebra
Summary:
This introduces the second layer of erlfmt_algebra with the lines API.
This API allows expressing any possible code layouts and the basic
operations of `new`, `combine`, and `flush` will be recurring through
the other layers.

The additional modifications in tests change the string generator to
maintain the invariant needed for lines that strings (that represent
single line) don't contain newlines.