Handle floats & chars in erlfmt_format
Summary:
This adds support for floats & chars to `erlfmt_format`.

Floats are normalised to downcase the exponent marker.

Chars are normalised for the escapes - explcit escapes in source are kept
as they were, except upcased - `$\xAA` is more readable than `$\xaa`.
Other escapes are canonicalised to either the short-form (e.g. $\n)
where available, or to $\xXX, where XX are hex digits.