Handle concat in erlfmt_format
Summary:
This introduces handling for the concat node in erlfmt_format.
For now the elements of concatentation are either printed in one line or
vertically without any indentation. I'm not sure this is the best way,
but I need to start somewhere.

There's also a new `erlfmt_algebra:document_fold/2` function added.