Handle operators in erlfmt_format
Summary:
This adds handling of operators to erlfmt_format.
Instead of relying always on just operator precedence, the formatter
will add some extra parentheses to improve readability. The choice is
based on what the Elixir formatter does. Operators are divided in 3 categories:

  * Always force parens on nested operators: bitwise and list operators
  * Forces parens when nested operator is from that group: boolean operators
  * Rely on precedence: the rest

Nested operator chains (like `A ++ B ++ C`) are always printed without
parens (unless precedence rules would require them). When breaking those
expressions over multiple lines, operators are left at the end of a line
and folling lines will be indented only once, so a line with `C` would
have the same indentation as a line with `B`.

To achive this a new combinator for algebra is introduced -
`document_single_line` returns only layouts that would fit in a single
line.