Replace cons + nil with list node in erlfmt_parse
Summary:
This is needed since othwerwise parser would produce exactly the same
AST for `[1, 2]` and `[1 | [2 | []]]`. There was a consensus in the team
that explicit consing syntax shouldn't be converted to list syntax by
the formatter.
The representation of the list node where the last element can be a
special `cons` node is taken from Elixir, where the explicit consing
syntax is represented simply by a `|` operator. I chose to use a special
node instead of an operator node since the formatting rules will be
different for them.