Handle lists in erlfmt_format
Summary:
Add list handling to erlfmt_format. Lists are either formatted with all
elements on the same line or with one element per line and wrapping `[`
and `]` on separate lines. If there's an explicit cons `|` used, it's
treated as a single element printed either on a single line or over two
lines with `|` at the beginning of the second line (this is the same as Elixir), e.g.

```
[
    11,
    22
    | 33
]
```