Handle map creation in erlfmt_format
Summary:
This adds support for map creation expressions in erlfmt_format. They
are printed the same as other containers, while `=>` and `:=` are treated
similar to operators and always have spaces around them. The value is
either printed in one line on a newline is inserted after the operator
and the value is printed nested.