Handle fun expressions in formatter
Summary:
This is probably the hardest choice in how to format something so far.
There's a lot of possiblities and probably everybody has some opinions.
The way we format funs here will ultimately also be the way we'll format
other clauses - in receive, case, catch, and functions.
Please see the tests for possible formattings.