Change AST for clauses
Summary:
This changes AST for the `clause` node in all the palces where it's used
and makes it have the same shape everywhere.

Additionally in case of `catch` clauses there used to be no way to
differentiate between `throw:_ ->` and `_ ->`. We might want to convert
one to the other, but I don't want to do this unconditionally just yet.