Normalize attributes
Summary:
Further normalize attributes. This includes removing special-handling of
multiple attributes, always keeping the attribute value as a list (which
allows easily representing 0-ary, 1-ary, and 2-ary attributes).
Additionally modify all attributes with special syntax to have a 2-tuple
values further descrbied in the docs.
This should make handling of the attributes in the formatter slightly
more regular and easier.
Additionally update all the specs in `erlfmt_parse` to be correct.