Change AST for types
Summary:
This changes the AST for types in erlfmt_parse.

The primary motivation for this is that types had a completely different AST than the rest of the language even though, they used similar syntax. In terms of erlfmt it meant that we either reimplement formatting for the second set of AST or we change the AST to be common. After consulting with @ilyaklyuchnikov, I chose to unify the AST representations.

My first attempt was to remove `type` production from the parser completely and just make expressions and types use the same productions. This proved impossible due to conflicts arising from similar, yet slightly different syntax for funs (`fun((ok) -> ok)` vs `fun ((ok)) -> ok end`).

The resulting parser is much more permissive than the original erlfmt,
I don't believe for this to be an issue.

The change in precedence of `catch` was necessary to resolve precedence issues in expression like `catch 1 :: foo()` which is not valid Erlang, but can arise in the current parser.