Handle record, spec & type attributes in erlfmt
Summary:
This adds handling for `-record`, `-type`, `-opaque`, `-spec`, and
`-callback` attributes.

I'm not 100% happy with the way they are indented by defaault, but it's
a start.

I copied over the precedence table function since we're now treating in-type and regular operators the same, while `erl_parse` has them separate.