Recomment forms after parsing
Summary:
This adds module erlfmt_recomment that inserts comments back into the
AST. They are not printed back yet.

Additionally there's a small normalisation of comments - a space is
always inserted after the sequence of `%` chars (this is based on what
Elixir does).

The comments are attributed to the earliest element that is present on a
line that is greater or equal to the line where the comment is present.
This means we don't allow comments "trailing" on the same line as an
expression - this follows Elixir and a principle for erlfmt that I
established that we won't to avoid "realignment" of things if their
names change - placing comments at the end of the line is inherently
prone to this. Those comments will be placed on the line above.

One problem of this recommenting approach is misplacing "trailing"
comments in containers, for example in:
```
[
    X
    % foo
],
Bar
```
The comment will be attached to `Bar`. This comes from the fact that the
parser does not track positions of closing elements (like `]`), only
positions of the opening elements (so `[`, `X` and `Bar` in this
example). This will be subsequently fixed.