Recomment forms after parsing
Summary:
This adds module erlfmt_recomment that inserts comments back into the
AST. They are not printed back yet.

Additionally there's a small normalisation of comments - a space is
always inserted after the sequence of `%` chars (this is based on what
Elixir does).

The comments are attributed to the earliest element that is present on a
line that is greater or equal to the line where the comment is present.
This means we don't allow comments "trailing" on the same line as an
expression - this follows Elixir and a principle for erlfmt that I
established that we won't to avoid "realignment" of things if their
names change - placing comments at the end of the line is inherently
prone to this. Those comments will be placed on the line above.

One problem of this recommenting approach is misplacing "trailing"
comments in containers, for example in:
```
[
    X
    % foo
],
Bar
```
The comment will be attached to `Bar`. This comes from the fact that the
parser does not track positions of closing elements (like `]`), only
positions of the opening elements (so `[`, `X` and `Bar` in this
example). This will be subsequently fixed.

Test Plan:
There are only smoke-tests of this recommenting - i.e. we parse the
whole source of WA and recomment it, but we don't check if the
recommenting was correct (yet).
This runs during `rebar3 ct erlfmt`

Reviewers: alanzimm, #erl2

Reviewed By: alanzimm

Differential Revision: https://phabricator.intern.facebook.com/D19177511

Tasks: T58741601

Tags: wa_dialyzer, erl2

Signature: 19177511:1576833724:ae848267dc752a4703cb1cf0aa577ce9fc577a45