Keep original parens
Summary:
According to our discussions, keep original parentheses in the output.
This is primarily achieved by maintaining an annotation with
information if the original expression was wrapped in parens or not.

Doing this, allowed to significantly simplify the code by getting rid of
everything related to properly maintaining precedence - we can just
rely on the original being correct.
In particular for not double-nesting operators, we can rely on the fact
that if the same operator is on the correct side wrt associativity, it
doesn't need parens.

Test Plan: `rebar3 ct erlfmt`

Reviewers: alanzimm, #erl2

Reviewed By: alanzimm

Subscribers: loscher

Differential Revision: https://phabricator.intern.facebook.com/D19311841

Tasks: T59333334

Tags: erl2, commitClose

Signature: 19311841:1578582734:0f839b08e16f9bc6c53e285b6d54d889e4cbfa19