Remove hacks around token text in formatter
Summary:
The formatter in couple places inspects the token text to decide what to do. This in general seems weird and probably won't be possible once we track more precise token/AST information.
The representation for the `define` attribute is slightly changed - instead of the first node being always a `macro_call` node, we use either a `var`, `atom` or `call` node, which closer reflects the original syntax and allows to avoid the text matching hack.

This is a part of the work to properly handle comments by precisely tracking expression ranges.