Track ranges in parser
Summary:
This adds full range tracking to all the AST nodes. Every AST node now has a `location` and `end_location` annotation that should span the entire range it represents. This should allow us to correctly place comments, but also, in the future, produce more meaningful error messages if this parser is used in other places.

This also modifies the ast of `unit:8` binary type to a `remote` node instead of a "raw" tuple, which was problematic to handle.