Fix location tracking for top-level forms.
Summary:
The location information for attributes and functions was slightly off. Fix this by only tracking the initial position inside the form itself and later re-adjust the position using the final dot token.

Additionally, change the AST for attributes to use a full `atom` node for name. I thought before this could be avoided, but having the location anno in there for the entire attribute node, but text anno for just the name made it kind of awkward. Using separate annos for the two restores sanity.