Make special typed nodes into regular :: operator
Summary:
When initially converting types into regular AST, I left the `::` operator
as a special `typed` node instead of the regular `op` node, given it
wasn't clear it wouldn't need some special processing.
Now, it's clear that this separation only creates extra code insted of
simplifying, so I think it would be beneficial to downgrade it to the
regular `op` node.