Buffer tokens in the scanner between forms
Summary:
In some cases, the scanner would misplace the comment tokens placing
them in the following form. For example in:
```
-foo(1). % comment
-bar(1).
```
the comment token would be emitted for the second form (since the
scanner stopped at the `.`). To combat this, the `erlfmt_scan` module
buffers forms and checks the front of the next form to verify it doesn't
contain `white_space`/`comment` tokens that should belong to this form.

Furthermore, the snaphot tests were extended by adding support of test
that change format before and after running the formatter.

Additionally simplify some specs & types in `erlfmt_parse` since dialyzer started complaining.