Correctly assign comments in between forms
Summary:
When a comment exists in between two forms, previously we would assign
it as a post-comment to the previous form, while it generally would be
expected to be a pre-comment of the following form.

Additionally fix comment insertion in functions - add pre-comments
directly to the `function` node instead of nesting them on the first
`clause` node.