Modify clause nodes to hold more metadata
Summary:
This changes the general structure of `clause` nodes from
`{clause, Anno, Name, Args, Guards, Body}` to
`{clause, Anno1, {call, Anno2, Name, Args}, Guards, Body}`.
This is in preparation for user-guided breaking of certain elements, in
particular keeping track of newlines following, for example, `(` or `->`
tokens. This would be problematic with the old structure, since there was
just one `Anno` for the whole clause and it wasn't possible to have separate
metadata for the whole clause and just arguments. With this split, this is now
possible.

There are some additional supporting changes documented in
`erl_parse.md`.