Allow force-breaking clauses
Summary:
This allows force-breaking clauses (functions, case, try, and similar)
into always inserting a newline after `->` even if all clauses would
otherwise fit on a single line. This is controlled by a newline always
following the first clause - this allows easily controlling the layout
of the entire function with just a single character.