Preserve empty lines in blocks
Summary:
Preserve empty lines in blocks inserted by users. Empty lines are often
used for grouping expressions inside longer functions, removing those
guidelines would be against the goal of the formatter of making the code
more readable. However, at most one empty line is allowed.

This is detected based on differences in location information in
following expressions. To make this effective we need to account for
comments in location information - comments attached to a node are now
treated as part of that node.

There are some additional changes to the algorithm to prevent empty
lines with extra whitespace.