Change format of tuples and binaries
Summary:
This uses, by default, a "fill" layout for tuples and binaries. This
means the layout engine will try to fit as many elements on a single
line as possible before breaking and repeating the process. The opening
and closing brackets (`{}` and `<<>>`) are "hugging" the contents.
If the tuple or binary was written down in an "exploded" form - with
every element on a spearate line, this will be preserved. Similar to
regular collections this is controlled by checking if there's a newline
between opening bracket and first element.

This required adding additional special-case for layout of -record
definitions that shouldn't be affected by that change and otherwise used
regular rules for laying out tuples.