Remove newline annotation
This was introduced to track newlines after certain tokens to implement the
"follow my lead" behaviour in lists & friends. But in the new implementation
in 0.2.0 we're using the location information directly to do that instead of
the newline annotations. Because of that they are not needed.

This effectively reverts 4246eb4287ca1c65ca2aea340ea5a182cbec8537.