Make erlfmt:format_string API-compatible with erlfmt:format_file
This modifies format_string to expose as close of an API and return types
as it makes sense to erlfmt:format_file. This should also allow correctly
handling syntax errors in the passed string instead of crashing.

Subsequently remove erlfmt:contains_pragma_string shrinking the size of the
public API.