Apply `exclude_files` to `files` absolute paths
When calling `erlfmt` with absolute paths for the files to format, the
excludes don't apply, even if the resolved paths of the excludes would
match the files to format.

Resolve the paths of the excludes based on the current working directory
and add them to the excludes.

This issue was observed when using the
[VS Code Erlang Formatter extension](https://marketplace.visualstudio.com/items?itemName=szTheory.erlang-formatter).
The extension invokes `rebar3 fmt` with the full path of the file to
format.